feat: modularize application using Flask Blueprints
