+++ templates/layout.html
@@ -73,10 +73,10 @@
                         <span class="navbar-text me-3">
                             Welcome, {{ current_user.username }}!
                         </span>
-                        <a href="{{ url_for('main.logout') }}" class="btn btn-outline-danger">Logout</a>
+                        <a href="{{ url_for('auth.logout') }}" class="btn btn-outline-danger">Logout</a>
                     {% else %}
-                        <a href="{{ url_for('main.login') }}" class="btn btn-outline-success me-2">Login</a>
-                        <a href="{{ url_for('main.signup') }}" class="btn btn-outline-primary">Sign Up</a>
+                        <a href="{{ url_for('auth.login') }}" class="btn btn-outline-success me-2">Login</a>
+                        <a href="{{ url_for('auth.signup') }}" class="btn btn-outline-primary">Sign Up</a>
                     {% endif %}
                 </div>
             </div>
